small fix in Dump
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@780 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/common/CBytestream.cpp
@@ -164,7 +164,7 @@ void CBytestream::Dump(void)
 			else
 				cout << "\\\\";
 		} else
-			cout << "\\" << hex << (uint)*it << dec;
+			cout << "\\" << hex << (uint)(uchar)*it << dec;
 
 		// Linebreak after 16 dumped bytes
 		if((i % 16) == 15)
